web/QReadWeb/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
frozenlist==1.7.0
geoip2==5.1.0
idna==3.10
maxminddb==2.7.0
multidict==6.5.0
pillow==11.2.1
propcache==0.3.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
qrcode==8.2
requests==2.32.4
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
urllib3==2.5.0
yarl==1.20.1
